ref 处理迁移代码无法编译通过的问题

+++ daxpay-single/daxpay-single-service/src/main/java/cn/bootx/platform/daxpay/core/channel/voucher/entity/Voucher.java
@@ -2,7 +2,7 @@ package cn.bootx.platform.daxpay.core.channel.voucher.entity;
 
 import cn.bootx.platform.common.core.function.EntityBaseFunction;
 import cn.bootx.platform.common.mybatisplus.base.MpBaseEntity;
-import cn.bootx.platform.daxpay.code.paymodel.VoucherCode;
+import cn.bootx.platform.daxpay.code.VoucherCode;
 import cn.bootx.platform.daxpay.core.channel.voucher.convert.VoucherConvert;
 import cn.bootx.platform.daxpay.dto.channel.voucher.VoucherDto;
 import cn.bootx.table.modify.annotation.DbColumn;
@@ -57,11 +57,11 @@ public class Voucher extends MpBaseEntity implements EntityBaseFunction<VoucherD
 
     /** 余额 */
     @DbComment("余额")
-    private BigDecimal balance;
+    private Integer balance;
 
     /** 预冻结额度 */
     @DbColumn(comment = "预冻结额度")
-    private BigDecimal freezeBalance;
+    private Integer freezeBalance;
 
 
     /** 是否长期有效 */
